+++ src/pages/DetailTodoPage.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 function DetailTodoPage() {
     return (
-        <div></div>
+        <div>상세화면</div>
     );
 }
